Fine-mapping and clustering SEs with SUR

The function feature selects with SuSiE and runs SUR with cluster-robust SEs

Usage

compute_finemap_regress(
  X,
  Y,
  Y.rep,
  R,
  id,
  nfolds,
  verbose = F,
  tx_names = NULL,
  coverage = 0.9,
  seed
)

Arguments

X

matrix, design matrix of SNP dosages

Y

matrix, matrix of G mean isoform expression across columns

Y.rep

matrix, matrix of G isoform expression with replicates

R

int, number of replicates

id

vector, vector of sample ids showing rep to id

nfolds

int, number of folds

verbose

logical

tx_names

vector, character vector of tx names - order of columns of Y

coverage

numeric, covarage of credible sets

seed

int, random seed

Value

isoTWAS model list