Conditional gene-level test

The function tests the gene association with trait conditional on isoforms

Usage

conditional_test(w_gene, w_iso, ld, z, gene_name, tx_name)

Arguments

w_gene

vector, weights for gene

w_iso

matrix, weights for significant isoforms

ld

matrix, LD matrix

z

vector, GWAS Z-scores

gene_name

character, gene name

tx_name

vector, character vector of tx names in order of columns of Y

Value

data frame of conditional Z-scores for all combinations of tx