Breiman and Friedman's curds and whey multivariate regression

The function trains the curds and whey multivariate regression

Usage

curds_whey(
  X,
  Y,
  family = "gaussian",
  alpha = 0.5,
  nfolds = 5,
  verbose,
  par = F,
  n.cores = NULL,
  tx_names = NULL
)

Arguments

X

matrix, design matrix of SNP dosages

Y

matrix, matrix of G isoform expression across columns

family

character, glmnet glm family

alpha

numeric, elastic net mixing parameter

nfolds

int, number of CV folds

verbose

logical

par

logical, uses mclapply to parallelize model fit

n.cores

int, number of parallel cores

tx_names

vector, character vector of tx names in order of columns of Y

Value

data frame of elastic net, lasso, and LMM based predictions