FOCUS-like gene-level finemapping

The function runs probabilistic finemapping on the gene/isoform-level

Usage

twas_finemap(
  res.df,
  z = "Z",
  Omega,
  V,
  max_enum = 3,
  cutoff = 0.9,
  verbose = F
)

Arguments

res.df

data.frame, TWAS results that at least has Z scores and transcript names

z

character, column name for Z scores

Omega

matrix, matrix of SNP to transcript effect in same order as res.df

V

matrix, LD matrix in the same order as Omega

max_enum

numeric, max number of causals in the credible set

cutoff

numeric, proportion of posterior explained by credible set

verbose

logical, verbose

Value

data frame of res.df with pips and whether isoform is in credible set